Added a second target which duplicates the original target however changing the base view to be a SCNView instead of a SKView.
The SKScene is then assigned to the SCNView.overlaySKScene property.

The purpose of this is to demonstrate whether the focus engine (which works fine for a pure SpriteKit app) works when the SpriteKit scene is being managed by a SceneKit view.

At the time I wrote this, it does not so far as I can tell.